moved nth_algebraic tests
