+++ sympy/solvers/ode/tests/test_ode.py
@@ -3694,136 +3694,6 @@ def test_nth_order_reducible():
     assert dsolve(eqn, f(x), hint='nth_order_reducible') in (sol, sols)
 
 
-def test_nth_algebraic():
-    eqn = Eq(Derivative(f(x), x), Derivative(g(x), x))
-    sol = Eq(f(x), C1 + g(x))
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), hint='nth_algebraic'), dsolve(eqn, f(x), hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-    eqn = (diff(f(x)) - x)*(diff(f(x)) + x)
-    sol = [Eq(f(x), C1 - x**2/2), Eq(f(x), C1 + x**2/2)]
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False)[0]
-    assert set(sol) == set(dsolve(eqn, f(x), hint='nth_algebraic'))
-    assert set(sol) == set(dsolve(eqn, f(x)))
-
-    eqn = (1 - sin(f(x))) * f(x).diff(x)
-    sol = Eq(f(x), C1)
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-    M, m, r, t = symbols('M m r t')
-    phi = Function('phi')
-    eqn = Eq(-M * phi(t).diff(t),
-             Rational(3, 2) * m * r**2 * phi(t).diff(t) * phi(t).diff(t,t))
-    solns = [Eq(phi(t), C1), Eq(phi(t), C1 + C2*t - M*t**2/(3*m*r**2))]
-    assert checkodesol(eqn, solns[0], order=2, solve_for_func=False)[0]
-    assert checkodesol(eqn, solns[1], order=2, solve_for_func=False)[0]
-    assert set(solns) == set(dsolve(eqn, phi(t), hint='nth_algebraic'))
-    assert set(solns) == set(dsolve(eqn, phi(t)))
-
-    eqn = f(x) * f(x).diff(x) * f(x).diff(x, x)
-    sol = Eq(f(x), C1 + C2*x)
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-    eqn = f(x) * f(x).diff(x) * f(x).diff(x, x) * (f(x) - 1)
-    sol = Eq(f(x), C1 + C2*x)
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-    eqn = f(x) * f(x).diff(x) * f(x).diff(x, x) * (f(x) - 1) * (f(x).diff(x) - x)
-    solns = [Eq(f(x), C1 + x**2/2), Eq(f(x), C1 + C2*x)]
-    assert checkodesol(eqn, solns[0], order=2, solve_for_func=False)[0]
-    assert checkodesol(eqn, solns[1], order=2, solve_for_func=False)[0]
-    assert set(solns) == set(dsolve(eqn, f(x), hint='nth_algebraic'))
-    assert set(solns) == set(dsolve(eqn, f(x)))
-
-
-def test_nth_algebraic_issue15999():
-    eqn = f(x).diff(x) - C1
-    sol = Eq(f(x), C1*x + C2) # Correct solution
-    assert checkodesol(eqn, sol, order=1, solve_for_func=False) == (True, 0)
-    assert dsolve(eqn, f(x), hint='nth_algebraic') == sol
-    assert dsolve(eqn, f(x)) == sol
-
-
-def test_nth_algebraic_redundant_solutions():
-    # This one has a redundant solution that should be removed
-    eqn = f(x)*f(x).diff(x)
-    soln = Eq(f(x), C1)
-    assert checkodesol(eqn, soln, order=1, solve_for_func=False)[0]
-    assert soln == dsolve(eqn, f(x), hint='nth_algebraic')
-    assert soln == dsolve(eqn, f(x))
-
-    # This has two integral solutions and no algebraic solutions
-    eqn = (diff(f(x)) - x)*(diff(f(x)) + x)
-    sol = [Eq(f(x), C1 - x**2/2), Eq(f(x), C1 + x**2/2)]
-    assert all(c[0] for c in checkodesol(eqn, sol, order=1, solve_for_func=False))
-    assert set(sol) == set(dsolve(eqn, f(x), hint='nth_algebraic'))
-    assert set(sol) == set(dsolve(eqn, f(x)))
-
-    eqn = f(x) + f(x)*f(x).diff(x)
-    solns = [Eq(f(x), 0),
-             Eq(f(x), C1 - x)]
-    assert all(c[0] for c in checkodesol(eqn, solns, order=1, solve_for_func=False))
-    assert set(solns) == set(dsolve(eqn, f(x)))
-
-    solns = [Eq(f(x), exp(x)),
-             Eq(f(x), C1*exp(C2*x))]
-    solns_final =  _remove_redundant_solutions(eqn, solns, 2, x)
-    assert solns_final == [Eq(f(x), C1*exp(C2*x))]
-
-    # This one needs a substitution f' = g.
-    eqn = -exp(x) + (x*Derivative(f(x), (x, 2)) + Derivative(f(x), x))/x
-    sol = Eq(f(x), C1 + C2*log(x) + exp(x) - Ei(x))
-    assert checkodesol(eqn, sol, order=2, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x))
-
-
-#
-# These tests can be combined with the above test if they get fixed
-# so that dsolve actually works in all these cases.
-#
-
-
-# prep = True breaks this
-def test_nth_algebraic_noprep1():
-    eqn = Derivative(x*f(x), x, x, x)
-    sol = Eq(f(x), (C1 + C2*x + C3*x**2) / x)
-    assert checkodesol(eqn, sol, order=3, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), prep=False, hint='nth_algebraic')
-
-
-@XFAIL
-def test_nth_algebraic_prep1():
-    eqn = Derivative(x*f(x), x, x, x)
-    sol = Eq(f(x), (C1 + C2*x + C3*x**2) / x)
-    assert checkodesol(eqn, sol, order=3, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), prep=True, hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-
-# prep = True breaks this
-def test_nth_algebraic_noprep2():
-    eqn = Eq(Derivative(x*Derivative(f(x), x), x)/x, exp(x))
-    sol = Eq(f(x), C1 + C2*log(x) + exp(x) - Ei(x))
-    assert checkodesol(eqn, sol, order=2, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), prep=False, hint='nth_algebraic')
-
-
-@XFAIL
-def test_nth_algebraic_prep2():
-    eqn = Eq(Derivative(x*Derivative(f(x), x), x)/x, exp(x))
-    sol = Eq(f(x), C1 + C2*log(x) + exp(x) - Ei(x))
-    assert checkodesol(eqn, sol, order=2, solve_for_func=False)[0]
-    assert sol == dsolve(eqn, f(x), prep=True, hint='nth_algebraic')
-    assert sol == dsolve(eqn, f(x))
-
-
 # Needs to be a way to know how to combine derivatives in the expression
 def test_factoring_ode():
     from sympy import Mul
